requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.36.0
diffusers>=0.24.0
accelerate>=0.25.0
gradio>=4.0.0
Pillow>=10.0.0
tqdm>=4.65.0
omegaconf>=2.3.0
python-dotenv>=1.0.0 